Add files via upload
In this Program the user can Login as Admin, Cashier or Manager. Program contains all the prices and quantities for the programs that this company sells. The Cashier can sell products and create the bill in the system. Manager can check the balance and the products that were sold, in case when there is no products left it is able to add these products in the program so the Cashier can create new bills for customers. 
In this program we can change the price for each product, we can add new product or delete them, but these operations are not possible by the cashier. The cashier can only build bills and print them.
Each cashier, manager and admin can login with their own password and username.
In this program I have used Java and CSS.
To make this program work I have used GUI, inheritance, inputs and outputs, interface, abstract class, files, javafx, etc.